[Study programme] Extract component OverallStatsTable
Let's rewrite this separately
